bump failed test
the pointer to a &'static str from one test was contaminating a &'static str from another. How is that possible? Perhaps rust placed the strings next to each other in memory, without null characters, and CString::from_raw read starting from the current string to the end of all strings?